requirements.txt
python-docx==1.1.0
openai==0.28
colorama==0.4.6
python-docx==1.1.0
python-dotenv